# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: HomeFinance API
  version: "1.0.0"
  summary: Autenticação, contas, categorias, lançamentos, importação e transferências
  description: |
    Contrato da API do HomeFinance. **Fonte de verdade** do que a API expõe
    (ADR-006). Nesta entrega os handlers são escritos à mão (ADR-011), e o
    teste `cmd/api/routes_test.go` garante que o conjunto de (método, caminho)
    daqui é exatamente o da tabela de rotas em `cmd/api/routes.go` — divergir
    quebra o build.

    Convenções:
      - JSON em `camelCase`; instantes em ISO 8601 UTC; **data civil**
        (`AAAA-MM-DD`) e **mês** (`AAAA-MM`) são tipos próprios, sem hora e sem
        fuso (ADR-019a).
      - Nome de campo vale na **caixa exata** declarada aqui, em todo corpo de
        requisição e em todo nível dele (emenda §13 da spec 0005): `CategoryId`
        ou `EMAIL` é 400, como qualquer campo fora do contrato. Um campo, um
        nome — vale também para o campo repetido no mesmo objeto, cujas
        ocorrências são todas conferidas.
      - Dinheiro **sempre** em centavos, `integer/int64`. Não existe `number` em
        campo monetário em lugar nenhum deste contrato: float em dinheiro é
        proibido em toda camada (ADR-003), e um `number` aqui seria a porta de
        entrada dele no frontend.
      - Erro sempre no formato único do schema `Error`.
      - Sessão em cookies `HttpOnly; Secure; SameSite=Strict; Path=/`, com
...

...
  license:
    name: Proprietário
servers:
...
paths:
...
  /imports:
    get:
      tags: [imports]
      operationId: listImports
      summary: Histórico dos últimos lotes de importação da casa
      description: |
        Até 50 lotes, do mais recente para o mais antigo. Lote `pending` traz
        `counts` (a análise por status) e `outcome` nulo; lote `committed` traz
        o contrário — as linhas de *staging* são apagadas **fisicamente** no
        commit (spec 0004 §3.6), e a análise por status deixa de existir.
      responses:
        "200":
          description: Lotes da casa
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportBatchList" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [imports]
      operationId: createImport
      summary: Envia o arquivo e ANALISA (fase 1 de 2 — nada é gravado ainda)
      description: |
        **Esta rota não cria lançamento nenhum.** Ela extrai, detecta o formato,
        parseia, deduplica e grava o resultado num *staging* (`import_batches` +
        `import_rows`) que vive **24 h**. Só `POST /imports/{id}/confirm` grava
        em `transactions` (ADR-024c).

        **A única rota com corpo em `multipart/form-data`** e a única com teto
        de corpo de **8 MiB** — exceção declarada por caminho exato, sem prefixo
        e sem curinga; todas as outras rotas continuam em 1 MiB (spec 0004
        §6.4). `Content-Type: application/json` aqui é 415.

        Parte com nome fora da allowlist de quatro é **400**, nunca ignorada em
        silêncio. O tipo real do arquivo vem dos *magic bytes* do ZIP, nunca da
        extensão nem do `Content-Type` da parte.

        **A instituição da conta NÃO escolhe o parser** — quem escolhe é a
        detecção por cabeçalho. A `institution` da conta é trava de
        consistência: detectou `nubank` numa conta marcada `c6` → 422
        `IMPORT_TARGET_MISMATCH`, e nada é importado. A trava mais forte é a
        outra: fatura só entra em conta `credit_card`, e extrato só entra em
        conta que não seja cartão.

        Rate limit próprio: 10/h por casa e 20/h por IP — é a rota mais cara do
        sistema (inflate, parse e varredura de deduplicação).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/CreateImportRequest" }
            encoding:
              file:
                contentType: text/csv, application/zip, application/octet-stream
              accountId: { contentType: text/plain }
              password: { contentType: text/plain }
              format: { contentType: text/plain }
      responses:
        "201":
          description: |
            Arquivo analisado e em *staging*. Nenhum lançamento foi criado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportBatch" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/ImportRejected" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-hf_access
      description: |
        JWT HS256 de vida curta (15 min). Sem o prefixo `__Host-` quando
        `COOKIE_SECURE=false` (apenas desenvolvimento).
...
  responses:
    NotFound:
      description: |
        Recurso nao encontrado. E a MESMA resposta para "nao existe" e para
        "existe, mas e de outra casa".
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationFailed:
      description: Dados inválidos
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
    Unauthenticated:
      description: Autenticação necessária
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    RateLimited:
      description: Muitas requisições
      headers:
        Retry-After:
          description: Segundos até a próxima tentativa permitida
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    PayloadTooLarge:
      description: |
        Corpo acima do limite DA ROTA. O teto global é 1 MiB; `POST /imports` é
        a única exceção, com 8 MiB, declarada por caminho exato — sem prefixo e
        sem curinga (spec 0004 §6.4).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    UnsupportedMediaType:
      description: |
        Content-Type diferente do exigido pela rota: `application/json` em todas,
        exceto `POST /imports`, que exige `multipart/form-data`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InternalError:
      description: Erro interno; nenhum detalhe é exposto
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ImportRejected:
      description: |
        O arquivo, o formato ou a conta de destino foram recusados por uma regra
        da importação. O `code` diz **o que a tela deve fazer** — ver
        `ImportErrorCode`. Nenhum lançamento foi criado.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ImportError" }
...
  schemas:
...
    ErrorCode:
      type: string
      description: |
        Conjunto fechado. O cliente decide o que fazer **pelo código**, nunca
        pelo texto da mensagem (D4 da spec 0003).

        Os seis `IMPORT_*` são de `POST /imports` e `POST /imports/{id}/confirm`,
        e existem porque cada um mapeia para uma ação diferente da interface — a
        tabela que explica cada um está em `ImportErrorCode`, que é o
        subconjunto devolvido por aquelas rotas.

        `KEYWORD_TAKEN` (409) é de `POST/PATCH /categories` e `/accounts`: a
        tela mostra "já está em *X*" com o `ownerId` — ação diferente do 400 de
        forma, por isso código próprio (D4 da spec 0003).

        `CONFLICT` (409) é de `POST /transfers/detect`: o estado mudou entre a
        leitura e a escrita da mesma operação, nada foi gravado, e a ação da
        tela é pedir a prévia de novo — diferente do 422 de teto, que não
        muda com uma nova tentativa (ADR-028).
      enum:
        - VALIDATION_FAILED
        - INVALID_CREDENTIALS
        - INVALID_CODE
        - EMAIL_NOT_VERIFIED
        - UNAUTHENTICATED
        - INVALID_SESSION
        - FORBIDDEN
        - NOT_FOUND
        - METHOD_NOT_ALLOWED
        - RATE_LIMITED
        - PAYLOAD_TOO_LARGE
        - UNSUPPORTED_MEDIA_TYPE
        - SERVICE_UNAVAILABLE
        - INTERNAL_ERROR
        - RESOURCE_IN_USE
        - KEYWORD_TAKEN
        - CONFLICT
        - IMPORT_PASSWORD_REQUIRED
        - IMPORT_PASSWORD_INVALID
        - IMPORT_FORMAT_UNKNOWN
        - IMPORT_FORMAT_AMBIGUOUS
        - IMPORT_TARGET_MISMATCH
        - IMPORT_FILE_REJECTED
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message:
              type: string
              description: Mensagem genérica em pt-BR; nunca contém detalhe interno.
    ValidationError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              const: VALIDATION_FAILED
            message: { type: string }
            fields:
              type: object
              description: Só existe em VALIDATION_FAILED.
              additionalProperties: { type: string }
...
    CivilDate:
      type: string
      description: |
        Data civil `AAAA-MM-DD` — sem hora e sem fuso. Não é um instante: o dia
        do calendário não muda quando o servidor muda de fuso (ADR-019a).
      pattern: "^\\d{4}-\\d{2}-\\d{2}$"
      examples: ["2026-09-12"]
...
    YearMonth:
      type: string
      description: |
        Mês civil `AAAA-MM`, sem dia, sem hora e sem fuso. É o "mês do app" —
        sempre **competência** (ADR-023c), nunca caixa, em listagem, painel,
        orçamento e relatório.
      pattern: "^\\d{4}-(0[1-9]|1[0-2])$"
      examples: ["2026-09"]
...
    Institution:
      type: string
      description: |
        Instituição financeira. Allowlist fechada: nada de texto livre aqui.

        **Ela não escolhe o parser da importação** — quem escolhe é a detecção
        por cabeçalho (ADR-024a). Se escolhesse, conta marcada `other` ficaria
        sem importação e conta marcada errado importaria com o parser errado
        **em silêncio**, invertendo o sinal de uma fatura inteira. Aqui ela é
        **trava de consistência**: instituição detectada diferente da da conta é
        422 `IMPORT_TARGET_MISMATCH`. `other` não trava nada — só perde a
        checagem.
      enum: [c6, nubank, other]
...
    ImportDocKind:
      type: string
      description: |
        Tipo de documento, detectado pelo cabeçalho. É ele que decide a trava
        mais forte da importação: `card_statement` só entra em conta
        `credit_card`, e `checking_statement` só entra em conta que **não** seja
        cartão.
      enum: [checking_statement, card_statement]
    ImportBatchStatus:
      type: string
      description: |
        `pending` é o único estado em que o lote pode ser confirmado ou
        descartado; `discarded` e `expired` respondem 404 no confirm. `expired`
        é posto pelo janitor quando as 24 h passam.
      enum: [pending, committed, discarded, expired]
    ImportEncoding:
      type: string
      description: |
        Codificação com que o CSV foi lido. Sem UTF-8 válido, a leitura cai para
        **Windows-1252** — e não Latin-1 puro, porque 0x80–0x9F carregam
        travessão, aspas curvas e o símbolo do euro, que bancos brasileiros usam.
      enum: [utf-8, windows-1252]
    ImportFormatID:
      type: string
      description: |
        Id do parser que ganhou a detecção — par (instituição × documento),
        versionado. Allowlist do registro: valor fora dela é 400.
      pattern: "^[a-z0-9]+(\\.[a-z0-9_]+){1,3}$"
      maxLength: 40
      examples: ["nubank.card_statement.v1"]
...
    ImportRowCounts:
      type: object
      additionalProperties: false
      required:
        [novo, repetido_no_arquivo, duplicado_exato, duplicado_excluido,
         possivel_duplicado, pagamento_de_fatura, transferencia_interna,
         transferencia_ja_registrada, rejeitado]
      description: |
        Quantas linhas caíram em cada status da análise. Existe enquanto o lote
        está `pending`; some quando as linhas de *staging* são apagadas.
      properties:
        novo: { type: integer, minimum: 0 }
        repetido_no_arquivo: { type: integer, minimum: 0 }
        duplicado_exato: { type: integer, minimum: 0 }
        duplicado_excluido: { type: integer, minimum: 0 }
        possivel_duplicado: { type: integer, minimum: 0 }
        pagamento_de_fatura: { type: integer, minimum: 0 }
        transferencia_interna: { type: integer, minimum: 0 }
        transferencia_ja_registrada: { type: integer, minimum: 0 }
        rejeitado: { type: integer, minimum: 0 }
    ImportOutcome:
      type: object
      additionalProperties: false
      required: [imported, restored, skipped, blocked, rejected, linked]
      description: |
        O que o confirm fez de fato. Sobrevive ao lote como histórico, mesmo
        depois que as linhas de *staging* são apagadas.
      properties:
        imported: { type: integer, minimum: 0 }
        restored:
          type: integer
          minimum: 0
          description: |
            Linhas cuja gêmea estava excluída logicamente e foram **restauradas**
            em vez de reinseridas (ADR-025f). A restauração altera apenas
            `deletedAt` e `updatedAt` — nenhum campo financeiro é tocado.
        skipped: { type: integer, minimum: 0 }
        blocked:
          type: integer
          minimum: 0
          description: |
            Linhas que o confirm recusou ao recalcular dentro da transação.
            Bloqueio **não** é falha do lote.
        rejected: { type: integer, minimum: 0 }
        linked:
          type: integer
          minimum: 0
          description: |
            Linhas vinculadas a uma perna de transferência que já existia
            (ação `link`). Não criaram lançamento e não entram em `imported`.
    StatementSuggestion:
      type: object
      additionalProperties: false
      required: [competenceMonth, closingDate, dueDate]
      description: |
        **Sugestão**, nunca decisão: vem de `statementClosingDay` e
        `statementDueDay` da conta quando configurados, e cai para uma inferência
        a partir da maior data do arquivo quando não. O usuário confirma no passo
        de revisão.

        O nome do arquivo entra como pista **apenas** aqui, e nunca como regra —
        nome de arquivo é entrada do cliente, e uma fatura arquivada no mês errado
        estragaria a competência de dezenas de linhas.
      properties:
        competenceMonth: { $ref: "#/components/schemas/YearMonth" }
        closingDate: { $ref: "#/components/schemas/CivilDate" }
        dueDate: { $ref: "#/components/schemas/CivilDate" }
    ImportBatch:
      type: object
      additionalProperties: false
      required:
        [id, status, accountId, institution, docKind, formatId, fileName, encoding,
         rowCount, minDate, maxDate, counts, outcome, statementSuggestion,
         sameContentImportedAt, expiresAt, createdAt, committedAt]
      properties:
        id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/ImportBatchStatus" }
        accountId: { type: string, format: uuid }
        institution:
          allOf:
            - $ref: "#/components/schemas/Institution"
          description: |
            Instituição **detectada no arquivo**, não a da conta. Nunca é
            `other`: um lote só existe quando um parser reconheceu o cabeçalho.
        docKind: { $ref: "#/components/schemas/ImportDocKind" }
        formatId: { $ref: "#/components/schemas/ImportFormatID" }
        fileName:
          type: string
          maxLength: 200
          description: |
            **Sanitizado**: sem separador de caminho, sem `..` e sem caractere de
            controle. É conteúdo do usuário e é exibido na tela, então passa
            pelas mesmas regras da descrição.
        encoding: { $ref: "#/components/schemas/ImportEncoding" }
        rowCount:
          type: integer
          minimum: 0
          description: Linhas de dados do arquivo. Teto de 10.000.
        minDate: { $ref: "#/components/schemas/CivilDate" }
        maxDate: { $ref: "#/components/schemas/CivilDate" }
        counts:
          oneOf:
            - $ref: "#/components/schemas/ImportRowCounts"
            - type: "null"
          description: |
            Análise por status. Presente enquanto o lote está `pending`; **nulo**
            depois que as linhas de *staging* são apagadas (confirm ou descarte).
        outcome:
          oneOf:
            - $ref: "#/components/schemas/ImportOutcome"
            - type: "null"
          description: Preenchido quando o lote foi confirmado; nulo antes disso.
        statementSuggestion:
          oneOf:
            - $ref: "#/components/schemas/StatementSuggestion"
            - type: "null"
          description: |
            Só existe quando `docKind = card_statement`. Nulo em extrato.
        sameContentImportedAt:
          type: [string, "null"]
          format: date-time
          description: |
            **Aviso, nunca bloqueio** (ADR-025d): quando preenchido, este mesmo
            conteúdo já foi importado nesta data. O hash é do conteúdo
            normalizado (pós-extração, sem BOM, com quebra de linha em LF), nunca
            dos bytes do arquivo enviado — duas compactações do mesmo CSV diferem
            em timestamp, nível de compressão e vetor da cifra.

            Bloquear seria frouxo (um byte a mais deixa passar um arquivo 99%
            repetido) e rigoroso ao mesmo tempo (travaria quem reimporta de
            propósito para pegar três linhas novas). **Economiza tempo; não é
            defesa.**
        expiresAt:
          type: string
          format: date-time
          description: |
            `createdAt` mais 24 h. Passou disso, o janitor marca `expired` e apaga
            as linhas — dado financeiro parado numa tabela de *staging* tem prazo.
        createdAt: { type: string, format: date-time }
        committedAt:
          type: [string, "null"]
          format: date-time
    ImportBatchList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ImportBatch" }
...
    CreateImportRequest:
      type: object
      additionalProperties: false
      required: [file, accountId]
      description: |
        Corpo de `POST /imports`, em `multipart/form-data`. **No máximo quatro
        partes**, e parte com nome fora desta lista é 400 — não se ignora em
        silêncio (spec 0004 §6.7).
      properties:
        file:
          type: string
          format: binary
          description: |
            CSV solto ou ZIP com uma **única** entrada `.csv`. Máximo de 8 MiB.
            O tipo real é decidido pelos **magic bytes**, nunca pela extensão nem
            pelo `Content-Type` da parte. Nada é escrito em disco em momento
            algum — a extração é 100% em memória.
        accountId:
          type: string
          format: uuid
          description: |
            Conta de destino, DA CASA DO TOKEN. Conta de outra casa responde
            **404**; conta arquivada responde 422 `VALIDATION_FAILED` com
            `fields.accountId`.
        password:
          type: string
          maxLength: 128
          writeOnly: true
          description: |
            Senha do ZIP. Só faz sentido em arquivo cifrado. **Nunca** é gravada,
            logada, auditada, ecoada nem devolvida em resposta nenhuma — vive
            como bytes, é usada e zerada. No cliente: campo `type="password"`,
            `autocomplete="off"`, fora de `sessionStorage` e de `localStorage`,
            apagada do estado assim que a requisição sai, e **jamais** em query
            string.
        format:
          allOf:
            - $ref: "#/components/schemas/ImportFormatID"
          description: |
            Desempate manual, necessário **apenas** depois de um
            `IMPORT_FORMAT_AMBIGUOUS`: o cliente reenvia o arquivo com um dos
            candidatos que vieram em `fields.format`. Fora da allowlist do
            registro é 400.
...
    ImportErrorCode:
      type: string
      description: |
        Subconjunto de `ErrorCode` que `POST /imports` e
        `POST /imports/{id}/confirm` podem devolver em **422**.

        Cada código existe porque mapeia para **uma ação diferente da interface**
        — é essa correspondência, e nada mais, que justifica seis códigos em vez
        de um. Colapsá-los em `VALIDATION_FAILED` obrigaria o frontend a
        interpretar texto em português, que é exatamente o que a D4 da spec 0003
        recusou. **Quem for "simplificar" isto depois: a tabela abaixo é o
        motivo.**

        | Código | Quando | O que a tela FAZ |
        |---|---|---|
        | `IMPORT_PASSWORD_REQUIRED` | ZIP com o bit de cifra ligado e sem senha | **abre o campo de senha** e explica que a do C6 é o CPF só com números — não diz "dados inválidos" |
        | `IMPORT_PASSWORD_INVALID` | o CRC não bateu depois de decifrar | **mantém o campo de senha e pede de novo**, sem perder o arquivo já escolhido |
        | `IMPORT_FORMAT_UNKNOWN` | nenhum parser reconheceu o cabeçalho | **lista os formatos suportados** — é a resposta para arquivo do C6 enquanto os parsers dele não existem |
        | `IMPORT_FORMAT_AMBIGUOUS` | mais de um parser casou | **mostra um seletor** com os candidatos, que vêm em `fields.format`, e reenvia com `format` explícito |
        | `IMPORT_TARGET_MISMATCH` | instituição diferente da da conta, ou fatura em conta que não é cartão | manda **trocar a conta de destino**, não o arquivo |
        | `IMPORT_FILE_REJECTED` | limites: tamanho, nº de linhas, razão de compressão, entradas no ZIP, janela de datas larga demais, arquivo vazio, mais de 20% de linhas rejeitadas | **explica qual limite estourou** (`fields` diz qual) e sugere importar por período menor. Um código só para todos os limites evita inflar a enum |

        `VALIDATION_FAILED` continua existindo para o resto — conta arquivada
        como destino, por exemplo, responde 422 `VALIDATION_FAILED` com
        `fields.accountId`, seguindo o padrão já existente.
      enum:
        - IMPORT_PASSWORD_REQUIRED
        - IMPORT_PASSWORD_INVALID
        - IMPORT_FORMAT_UNKNOWN
        - IMPORT_FORMAT_AMBIGUOUS
        - IMPORT_TARGET_MISMATCH
        - IMPORT_FILE_REJECTED
        - VALIDATION_FAILED
    ImportError:
      type: object
      additionalProperties: false
      required: [error]
      description: |
        Envelope de erro da importação. Mesma forma do `Error` de sempre, com o
        `code` restrito ao conjunto que a tela de importação sabe tratar e com
        `fields` opcional.
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ImportErrorCode" }
            message:
              type: string
              description: |
                Mensagem genérica em pt-BR. Nunca contém detalhe interno, trecho
                do arquivo, linha crua nem — jamais — a senha do ZIP.
            fields:
              type: object
              additionalProperties: { type: string }
              description: |
                Contexto legível por máquina. Em `IMPORT_FORMAT_AMBIGUOUS`,
                `fields.format` traz os candidatos; em `IMPORT_FILE_REJECTED`,
                diz qual limite estourou. Nunca carrega conteúdo do arquivo.
security:
  - sessionCookie: []